[ci] Add timeout for bootstrap and cleanup e2e cluster jobs
